backend/requirements.txt
bcrypt==4.0.1
fastapi==0.94.0
gunicorn==20.1.0
passlib==1.7.4
pytest==7.2.2
python-jose==3.3.0
requests==2.28.2
tortoise-orm==0.19.3
uvicorn==0.21.0
websockets==10.4
email-validator==2.0.0.post2
aiosmtplib==2.0.2
python-dotenv==1.0.0
pyOpenSSL==23.2.0
APScheduler==3.10.1
aiomysql==0.1.1
tenacity==8.2.2
